Isolate application graph test fixtures (#1957)
