+++ src/test-utils/application-graph.ts
@@ -1,0 +1,379 @@
+import { type Mock, vi } from 'vitest';
+
+const hoistedApplicationGraphMocks = vi.hoisted(() => ({
+  computeKanbanColumn: vi.fn(),
+  computePendingRequestType: vi.fn(),
+  deriveWorkspaceFlowStateFromWorkspace: vi.fn(),
+}));
+
+export interface ApplicationGraphMocks {
+  computeKanbanColumn: Mock<(...args: unknown[]) => unknown>;
+  computePendingRequestType: Mock<(...args: unknown[]) => unknown>;
+  deriveWorkspaceFlowStateFromWorkspace: Mock<(...args: unknown[]) => unknown>;
+}
+
+export const applicationGraphMocks: ApplicationGraphMocks = hoistedApplicationGraphMocks;
+
+vi.mock('@/backend/db', () => ({ prisma: { $disconnect: vi.fn() } }));
+vi.mock('@/backend/interceptors', () => ({
+  registerInterceptors: vi.fn(),
+  startInterceptors: vi.fn(),
+  stopInterceptors: vi.fn(),
+}));
+vi.mock('@/backend/orchestration/cli-health.service', () => ({ cliHealthService: {} }));
+vi.mock('@/backend/orchestration/data-backup.service', () => ({ dataBackupService: {} }));
+vi.mock('@/backend/orchestration/decision-log-query.service', () => ({
+  decisionLogQueryService: {},
+}));
+vi.mock('@/backend/orchestration/domain-bridges.orchestrator', () => ({
+  configureDomainBridges: vi.fn(),
+}));
+vi.mock('@/backend/orchestration/event-collector.orchestrator', () => ({
+  createEventCollectorOrchestrator: vi.fn(() => ({ configure: vi.fn(), stop: vi.fn() })),
+}));
+vi.mock('@/backend/orchestration/health.service', () => ({ healthService: {} }));
+vi.mock('@/backend/orchestration/reconciliation.service', () => ({ reconciliationService: {} }));
+vi.mock('@/backend/orchestration/scheduler.service', () => ({ schedulerService: {} }));
+vi.mock('@/backend/orchestration/snapshot-reconciliation.orchestrator', () => ({
+  SnapshotReconciliationService: class {
+    configure = vi.fn();
+    start = vi.fn();
+    stop = vi.fn(async () => undefined);
+  },
+}));
+vi.mock('@/backend/orchestration/workspace-archive.orchestrator', () => ({
+  archiveWorkspace: vi.fn(),
+  cleanupWorkspaceRuntimeResources: vi.fn(),
+  recoverStaleArchivingWorkspaces: vi.fn(),
+}));
+vi.mock('@/backend/orchestration/workspace-children.orchestrator', () => ({
+  createChildWorkspace: vi.fn(),
+  fireLifecycleNotification: vi.fn(),
+  persistChildNotification: vi.fn(),
+  persistParentNotification: vi.fn(),
+}));
+vi.mock('@/backend/orchestration/workspace-init.orchestrator', () => ({
+  initializeWorkspaceWorktree: vi.fn(),
+  retryQueuedDispatchAfterWorkspaceReady: vi.fn(),
+}));
+vi.mock('@/backend/orchestration/workspace-init-script-pipeline', () => ({
+  executeStartupScriptPipeline: vi.fn(),
+}));
+vi.mock('@/backend/prompts/quick-actions', () => ({
+  getQuickAction: vi.fn(),
+  listQuickActions: vi.fn(),
+}));
+vi.mock('@/backend/services/auto-iteration', () => ({
+  autoIterationService: {},
+  insightsService: {},
+  logbookService: {},
+}));
+vi.mock('@/backend/services/config.service', () => ({ configService: {} }));
+vi.mock('@/backend/services/crypto.service', () => ({ cryptoService: {} }));
+vi.mock('@/backend/services/factory-config.service', () => ({
+  FactoryConfigService: { readConfig: vi.fn() },
+}));
+vi.mock('@/backend/services/git-clone.service', () => ({ gitCloneService: {} }));
+vi.mock('@/backend/services/github', () => ({
+  githubCLIService: {},
+  prFetchRegistry: {},
+  prSnapshotService: {},
+}));
+vi.mock('@/backend/services/linear', () => ({ linearClientService: {} }));
+vi.mock('@/backend/services/logger.service', () => ({
+  createLogger: vi.fn(() => ({
+    debug: vi.fn(),
+    info: vi.fn(),
+    warn: vi.fn(),
+    error: vi.fn(),
+  })),
+  getLogFilePath: vi.fn(),
+}));
+vi.mock('@/backend/services/periodic-task', () => ({
+  periodicTaskAccessor: {},
+  periodicTaskService: {},
+}));
+vi.mock('@/backend/services/port.service', () => ({ findAvailablePort: vi.fn() }));
+vi.mock('@/backend/services/ratchet', () => ({ fixerSessionService: {}, ratchetService: {} }));
+vi.mock('@/backend/services/rate-limiter.service', () => ({ rateLimiter: {} }));
+vi.mock('@/backend/services/run-script', () => ({
+  createRunScriptService: vi.fn(() => ({})),
+  runScriptStateMachine: {},
+  startupScriptService: {},
+}));
+vi.mock('@/backend/services/run-script-config-persistence.service', () => ({
+  runScriptConfigPersistenceService: {},
+}));
+vi.mock('@/backend/services/server-instance.service', () => ({
+  createServerInstanceService: vi.fn(() => ({})),
+}));
+vi.mock('@/backend/services/session', () => ({
+  acpRuntimeManager: {},
+  acpTraceLogger: {},
+  chatEventForwarderService: {},
+  chatMessageHandlerService: {},
+  fetchCodexModelCatalogFromAppServer: vi.fn(),
+  sessionDataService: {},
+  sessionDomainService: {},
+  sessionEventBus: {},
+  sessionFileLogger: {},
+  sessionProviderResolverService: {},
+  sessionService: {},
+}));
+vi.mock('@/backend/services/terminal', () => ({ terminalService: {} }));
+vi.mock('@/backend/services/workspace', () => ({
+  computeKanbanColumn: (...args: unknown[]) => applicationGraphMocks.computeKanbanColumn(...args),
+  computePendingRequestType: (...args: unknown[]) =>
+    applicationGraphMocks.computePendingRequestType(...args),
+  deriveWorkspaceFlowStateFromWorkspace: (...args: unknown[]) =>
+    applicationGraphMocks.deriveWorkspaceFlowStateFromWorkspace(...args),
+  getWorkspaceInitPolicy: vi.fn(),
+  kanbanStateService: {},
+  projectManagementService: {},
+  userSettingsQueryService: {},
+  WorkspaceCreationService: class {},
+  workspaceAccessor: {},
+  workspaceActivityService: {},
+  workspaceDataService: {},
+  workspaceNotificationAccessor: {},
+  workspaceQueryService: {},
+  workspaceStateMachine: {},
+  worktreeLifecycleService: {},
+}));
+vi.mock('@/backend/services/workspace-snapshot-store.service', () => ({
+  workspaceSnapshotStore: {},
+}));
+
+import type {
+  ApplicationDependencies,
+  ApplicationLifecycle,
+  ApplicationServices,
+} from '@/backend/app-context';
+import { prisma } from '@/backend/db';
+import { registerInterceptors, startInterceptors, stopInterceptors } from '@/backend/interceptors';
+import { cliHealthService } from '@/backend/orchestration/cli-health.service';
+import { dataBackupService } from '@/backend/orchestration/data-backup.service';
+import { decisionLogQueryService } from '@/backend/orchestration/decision-log-query.service';
+import type { configureDomainBridges } from '@/backend/orchestration/domain-bridges.orchestrator';
+import { createEventCollectorOrchestrator } from '@/backend/orchestration/event-collector.orchestrator';
+import { healthService } from '@/backend/orchestration/health.service';
+import { reconciliationService } from '@/backend/orchestration/reconciliation.service';
+import { schedulerService } from '@/backend/orchestration/scheduler.service';
+import { SnapshotReconciliationService } from '@/backend/orchestration/snapshot-reconciliation.orchestrator';
+import {
+  archiveWorkspace,
+  cleanupWorkspaceRuntimeResources,
+  recoverStaleArchivingWorkspaces,
+} from '@/backend/orchestration/workspace-archive.orchestrator';
+import {
+  createChildWorkspace,
+  fireLifecycleNotification,
+  persistChildNotification,
+  persistParentNotification,
+} from '@/backend/orchestration/workspace-children.orchestrator';
+import {
+  initializeWorkspaceWorktree,
+  retryQueuedDispatchAfterWorkspaceReady,
+} from '@/backend/orchestration/workspace-init.orchestrator';
+import { executeStartupScriptPipeline } from '@/backend/orchestration/workspace-init-script-pipeline';
+import { getQuickAction, listQuickActions } from '@/backend/prompts/quick-actions';
+import {
+  autoIterationService,
+  insightsService,
+  logbookService,
+} from '@/backend/services/auto-iteration';
+import { configService } from '@/backend/services/config.service';
+import { cryptoService } from '@/backend/services/crypto.service';
+import { FactoryConfigService } from '@/backend/services/factory-config.service';
+import { gitCloneService } from '@/backend/services/git-clone.service';
+import { githubCLIService, prFetchRegistry, prSnapshotService } from '@/backend/services/github';
+import { linearClientService } from '@/backend/services/linear';
+import { createLogger, getLogFilePath } from '@/backend/services/logger.service';
+import { periodicTaskAccessor, periodicTaskService } from '@/backend/services/periodic-task';
+import { findAvailablePort } from '@/backend/services/port.service';
+import { fixerSessionService, ratchetService } from '@/backend/services/ratchet';
+import { rateLimiter } from '@/backend/services/rate-limiter.service';
+import {
+  createRunScriptService,
+  runScriptStateMachine,
+  startupScriptService,
+} from '@/backend/services/run-script';
+import { runScriptConfigPersistenceService } from '@/backend/services/run-script-config-persistence.service';
+import { createServerInstanceService } from '@/backend/services/server-instance.service';
+import {
+  acpRuntimeManager,
+  acpTraceLogger,
+  chatEventForwarderService,
+  chatMessageHandlerService,
+  fetchCodexModelCatalogFromAppServer,
+  sessionDataService,
+  sessionDomainService,
+  sessionEventBus,
+  sessionFileLogger,
+  sessionProviderResolverService,
+  sessionService,
+} from '@/backend/services/session';
+import { terminalService } from '@/backend/services/terminal';
+import {
+  getWorkspaceInitPolicy,
+  kanbanStateService,
+  projectManagementService,
+  userSettingsQueryService,
+  workspaceAccessor,
+  workspaceActivityService,
+  workspaceDataService,
+  workspaceNotificationAccessor,
+  workspaceQueryService,
+  workspaceStateMachine,
+  worktreeLifecycleService,
+} from '@/backend/services/workspace';
+import { workspaceSnapshotStore } from '@/backend/services/workspace-snapshot-store.service';
+
+const fakeSystemConfig = {
+  baseDir: '/tmp/factory-factory',
+  worktreeBaseDir: '/tmp/factory-factory/worktrees',
+  reposDir: '/tmp/factory-factory/repos',
+  debugLogDir: '/tmp/factory-factory/debug',
+  acpTraceLogsPath: '/tmp/factory-factory/debug/acp-events',
+  claudeConfigDir: '/tmp/.claude',
+  wsLogsPath: '/tmp/factory-factory/ws-logs',
+  backendPort: 3001,
+  nodeEnv: 'test' as const,
+  shellPath: '/bin/sh',
+  databasePath: '/tmp/factory-factory/data.db',
+  defaultSessionProfile: {
+    model: 'test',
+    permissionMode: 'strict' as const,
+    maxTokens: 1000,
+    temperature: 0,
+  },
+  healthCheckIntervalMs: 1000,
+  maxSessionsPerWorkspace: 2,
+  logger: { level: 'debug' as const, prettyPrint: false, serviceName: 'test' },
+  rateLimiter: {
+    claudeRequestsPerMinute: 10,
+    claudeRequestsPerHour: 100,
+    maxQueueSize: 10,
+    queueTimeoutMs: 1000,
+  },
+  notification: { soundEnabled: false, pushEnabled: false },
+  cors: { allowedOrigins: [] },
+  debug: { chatWebSocket: false },
+  acpStartupTimeoutMs: 4321,
+  acpTraceLogsEnabled: false,
+  wsLogsEnabled: false,
+  runScriptProxyEnabled: false,
+  compression: { enabled: false },
+  branchRenameMessageThreshold: 5,
+  appVersion: 'test',
+};
+
+export interface FakeApplicationGraph extends ApplicationDependencies {
+  config: typeof fakeSystemConfig;
+}
+
+export function createFakeApplicationGraph(label = 'test'): FakeApplicationGraph {
+  const graphAcpRuntimeManager = Object.assign({}, acpRuntimeManager, {
+    setAcpStartupTimeoutMs: vi.fn(),
+    configureEnvironment: vi.fn(),
+  }) satisfies typeof acpRuntimeManager;
+  const graphConfigService = Object.assign({}, configService, {
+    getAcpStartupTimeoutMs: vi.fn(() => 4321),
+    isProduction: vi.fn(() => false),
+    getChildProcessEnv: vi.fn(() => ({ APPLICATION_LABEL: label })),
+    getSystemConfig: vi.fn(() => fakeSystemConfig),
+    getWorktreeBaseDir: vi.fn(() => fakeSystemConfig.worktreeBaseDir),
+    getMaxSessionsPerWorkspace: vi.fn(() => fakeSystemConfig.maxSessionsPerWorkspace),
+    getCorsConfig: vi.fn(() => fakeSystemConfig.cors),
+  }) satisfies typeof configService;
+  const graphSessionService = Object.assign({}, sessionService, {
+    getRuntimeSnapshot: vi.fn(),
+  }) satisfies typeof sessionService;
+  const graphChatEventForwarderService = Object.assign({}, chatEventForwarderService, {
+    getAllPendingRequests: vi.fn(() => new Map()),
+  }) satisfies typeof chatEventForwarderService;
+
+  const services = {
+    acpRuntimeManager: graphAcpRuntimeManager,
+    acpTraceLogger,
+    autoIterationService,
+    archiveWorkspace,
+    chatEventForwarderService: graphChatEventForwarderService,
+    chatMessageHandlerService,
+    cleanupWorkspaceRuntimeResources,
+    cliHealthService,
+    configService: graphConfigService,
+    createChildWorkspace,
+    createLogger,
+    createWorkspaceCreationService: () => ({ create: vi.fn() }),
+    cryptoService,
+    dataBackupService,
+    decisionLogQueryService,
+    executeStartupScriptPipeline,
+    factoryConfigService: FactoryConfigService,
+    fetchCodexModelCatalogFromAppServer,
+    findAvailablePort,
+    fireLifecycleNotification,
+    fixerSessionService,
+    getLogFilePath,
+    getQuickAction,
+    getWorkspaceInitPolicy,
+    gitCloneService,
+    githubCLIService,
+    healthService,
+    initializeWorkspaceWorktree,
+    insightsService,
+    kanbanStateService,
+    linearClientService,
+    listQuickActions,
+    logbookService,
+    periodicTaskAccessor,
+    periodicTaskService,
+    persistChildNotification,
+    persistParentNotification,
+    prFetchRegistry,
+    prSnapshotService,
+    projectManagementService,
+    rateLimiter,
+    ratchetService,
+    reconciliationService,
+    retryQueuedDispatchAfterWorkspaceReady,
+    runScriptConfigPersistenceService,
+    runScriptService: createRunScriptService({ registerShutdownHandlers: false }),
+    runScriptStateMachine,
+    schedulerService,
+    serverInstanceService: createServerInstanceService(),
+    sessionDataService,
+    sessionDomainService,
+    sessionEventBus,
+    sessionFileLogger,
+    sessionProviderResolverService,
+    sessionService: graphSessionService,
+    startupScriptService,
+    terminalService,
+    userSettingsQueryService,
+    workspaceAccessor,
+    workspaceActivityService,
+    workspaceDataService,
+    workspaceNotificationAccessor,
+    workspaceQueryService,
+    workspaceSnapshotStore,
+    workspaceStateMachine,
+    worktreeLifecycleService,
+  } satisfies ApplicationServices;
+
+  const lifecycle = {
+    database: prisma,
+    interceptors: {
+      register: registerInterceptors,
+      start: startInterceptors,
+      stop: stopInterceptors,
+    },
+    wireDomainBridges: vi.fn<typeof configureDomainBridges>(),
+    eventCollector: createEventCollectorOrchestrator(),
+    snapshotReconciliation: new SnapshotReconciliationService(),
+    recoverStaleArchivingWorkspaces,
+  } satisfies ApplicationLifecycle;
+
+  return { services, lifecycle, config: fakeSystemConfig };
+}
